Update to latest CC
The main change is the fact that the ScalarField now has an offset
system baked into it, meaning APIs work with double but internally
floats are stored

This has the consequence of making the asArray to convert to a numpy
array view of the scalar field impossible.

As a replacement a custom class has been made, its very incomplete
as it supports only assignments and read bot no operations like
+,-,==, etc.

So now users will be forced to use toArray to copy the values into a
numpy array, work with them then copy back into the scalar field when
needed

+++ tests/scripts/normal_distribution.py
@@ -20,5 +20,5 @@ cloud = CC.clouds()[0].pc
 d.computeParameters(cloud)
 
 d.setParameters(1.1, 0.6)
-assert d.getMu() == float(np.float32(1.1))
-assert d.getSigma2() == float(np.float32(0.6))
+assert d.getMu() == 1.1
+assert d.getSigma2() == 0.6
